+++ live-demos/examples/animated-subscribe-button.tsx
@@ -1,0 +1,9 @@
+import { AnimatedSubscribeButton } from '../vendor/magic-ui/registry/magicui/animated-subscribe-button';
+export default function Demo() {
+  return (
+    <AnimatedSubscribeButton>
+      <span>Subscribe</span>
+      <span>Subscribed</span>
+    </AnimatedSubscribeButton>
+  );
+}